Crate bp_polkadot_core

Re-exports

pub use time_units::*;

Modules

time_units

Human readable time units defined in terms of number of blocks.

Structs

AccountIdConverter

Convert a 256-bit hash into an AccountId.

BlockLength

All Polkadot-like chains have maximal block size set to 5MB.

BlockWeights

All Polkadot-like chains have the same block weights.

ExtrinsicBaseWeight

Executing 10,000 System remarks (no-op) txs takes ~1.26 seconds -> ~125 µs per tx

Perbill

A fixed point representation of a number in the range [0, 1].

PolkadotLike

Polkadot-like chain.

SignedExtensions

A simplified version of signed extensions meant for producing signed transactions and signed payload in the client code.

Constants

ADDITIONAL_MESSAGE_BYTE_DELIVERY_WEIGHT

Increase of delivery transaction weight on Polkadot-like chain with every additional message byte.

AVERAGE_ON_INITIALIZE_RATIO

All Polkadot-like chains assume that an on-initialize consumes 1 percent of the weight on average, hence a single extrinsic will not be allowed to consume more than AvailableBlockRatio - 1 percent.

DEFAULT_MESSAGE_DELIVERY_TX_WEIGHT

Weight of single regular message delivery transaction on Polkadot-like chain.

EXTRA_STORAGE_PROOF_SIZE

Number of extra bytes (excluding size of storage value itself) of storage proof, built at Polkadot-like chain. This mostly depends on number of entries in the storage trie. Some reserve is reserved to account future chain growth.

MAXIMAL_ENCODED_ACCOUNT_ID_SIZE

Maximal size (in bytes) of encoded (using Encode::encode()) account id.

MAXIMUM_BLOCK_WEIGHT

All Polkadot-like chains allow 2 seconds of compute with a 6-second average block time.

MAX_MESSAGES_IN_DELIVERY_TRANSACTION

Maximal number of messages in single delivery transaction.

MAX_SINGLE_MESSAGE_DELIVERY_CONFIRMATION_TX_WEIGHT

Maximal weight of single message delivery confirmation transaction on Polkadot-like chain.

MAX_UNCONFIRMED_MESSAGES_AT_INBOUND_LANE

Maximal number of unconfirmed messages at inbound lane.

MAX_UNREWARDED_RELAYER_ENTRIES_AT_INBOUND_LANE

Maximal number of unrewarded relayer entries at inbound lane.

PAY_INBOUND_DISPATCH_FEE_WEIGHT

Weight of pay-dispatch-fee operation for inbound messages at Polkadot-like chain.

TX_EXTRA_BYTES

Maximal number of bytes, included in the signed Polkadot-like transaction apart from the encoded call itself.

Traits

Convert

Extensible conversion trait. Generic over both source and destination types.

Parameter

A type that can be used as a parameter in a dispatchable function.

Functions

account_info_storage_key

Return a storage key for account data.

max_extrinsic_size

Get the maximum length in bytes that a Normal extrinsic on the Polkadot-like chain requires.

max_extrinsic_weight

Get the maximum weight (compute time) that a Normal extrinsic on the Polkadot-like chain can use.

Type Definitions

AccountAddress

Address of account on Polkadot-like chains.

AccountId

Id of account on Polkadot-like chains.

AccountPublic

Public key of account on Polkadot-like chains.

AdditionalSigned

Parameters which are part of the payload used to produce transaction signature, but don’t end up in the transaction itself (i.e. inherent part of the runtime).

Address

Account address, used by the Polkadot-like chain.

Balance

The balance of an account on Polkadot-like chain.

Block

Block type of Polkadot-like chains.

BlockNumber

Block number type used in Polkadot-like chains.

Hash

Hash type used in Polkadot-like chains.

Hasher

The type of object that can produce hashes on Polkadot-like chains.

Hashing

Hashing type.

Header

The header type used by Polkadot-like chains.

Index

Account Index (a.k.a. nonce).

Nonce

Index of a transaction on the Polkadot-like chains.

Signature

Signature type used by Polkadot-like chains.

SignedBlock

Polkadot-like block signed with a Justification.

SignedExtra

A type of the data encoded as part of the transaction.

UncheckedExtrinsic

Unchecked Extrinsic type.